WIP: Int ID移行ビルドエラー修正中
- InventoryProgressService: xxxIndex → xxxId
- ShopProgressService: itemId型修正
- ProgressService.ItemSale: 全面修正
- ProgressService.ExplorationRun: dungeonId → UInt8
- その他の型変換修正

残りエラー: 約30件
- ItemSortOrderCalculator.swift
- ProgressService.ExplorationRuntime.swift
- ProgressService.StoryUnlocks.swift
- ItemSynthesisProgressService.swift

+++ Epika/Progress/Application/Services/ShopProgressService.swift
@@ -196,7 +196,7 @@ actor ShopProgressService {
         try saveIfNeeded(context)
 
         _ = try await inventoryService.addItem(itemId: target.definition.id,
-                                               quantity: UInt16(quantity),
+                                               quantity: quantity,
                                                storage: .playerItem)
         return playerSnapshot
     }
